+++ src/games/breakout/src/interfaces/interfaces.ts
@@ -1,0 +1,9 @@
+export interface IRectangleConstructor {
+  scene: Phaser.Scene;
+  x: number;
+  y: number;
+  width?: number;
+  height?: number;
+  fillColor?: number;
+  fillAlpha?: number;
+}